+++ src/Heist/Extra/Splices/Pandoc/Ctx.hs
@@ -32,33 +32,23 @@ data RenderCtx = RenderCtx
   , -- Custom render functions for AST nodes.
     blockSplice :: B.Block -> Maybe (HI.Splice Identity)
   , inlineSplice :: B.Inline -> Maybe (HI.Splice Identity)
+  , enableSyntaxHighlighting :: Bool
+  -- ^ Enable syntax highlighting for code blocks using skylighting
   }
 
 mkRenderCtx ::
   (Monad m) =>
-  Map Text Text ->
-  (RenderCtx -> B.Block -> Maybe (HI.Splice Identity)) ->
-  (RenderCtx -> B.Inline -> Maybe (HI.Splice Identity)) ->
-  H.HeistT Identity m RenderCtx
-mkRenderCtx classMap bS iS = do
-  node <- H.getParamNode
-  pure $
-    mkRenderCtxWith
-      node
-      classMap
-      bS
-      iS
-
-mkRenderCtxWith ::
-  X.Node ->
   -- | How to replace classes in Div and Span nodes.
   Map Text Text ->
   -- | Custom handling of AST block nodes
   (RenderCtx -> B.Block -> Maybe (HI.Splice Identity)) ->
   -- | Custom handling of AST inline nodes
   (RenderCtx -> B.Inline -> Maybe (HI.Splice Identity)) ->
-  RenderCtx
-mkRenderCtxWith node classMap bS iS = do
+  -- | Enable syntax highlighting for code blocks
+  Bool ->
+  H.HeistT Identity m RenderCtx
+mkRenderCtx classMap bS iS enableHighlighting = do
+  node <- H.getParamNode
   let ctx =
         RenderCtx
           (Just node)
@@ -67,11 +57,12 @@ mkRenderCtxWith node classMap bS iS = do
           classMap
           (bS ctx)
           (iS ctx)
-   in ctx
+          enableHighlighting
+   in pure ctx
 
 emptyRenderCtx :: RenderCtx
 emptyRenderCtx =
-  RenderCtx Nothing (const B.nullAttr) (const B.nullAttr) mempty (const Nothing) (const Nothing)
+  RenderCtx Nothing (const B.nullAttr) (const B.nullAttr) mempty (const Nothing) (const Nothing) False
 
 -- | Strip any custom splicing out of the given render context
 ctxSansCustomSplicing :: RenderCtx -> RenderCtx